Alterações Finais

+++ src/app/lista/page.tsx
@@ -1,10 +1,13 @@
 "use client";
 import { useState, useEffect } from "react";
-
+import Image from "next/image";
 // Interface para tipagem dos itens da lista
 interface ListItem {
   id: string;
-  text: string;
+  name: string;
+  resp: string;
+  date: string;
+  desc: string;
 }
 
 export default function Lista() {
@@ -18,36 +21,44 @@ export default function Lista() {
     }
   }, []);
 
-  // Remove item
   const handleDelete = (id: string) => {
-    const updatedItems = items.filter((item) => item.id !== id);
-
-    // Atualiza state e localStorage
-    setItems(updatedItems);
-    localStorage.setItem("listItems", JSON.stringify(updatedItems));
+    if (window.confirm("Tem certeza que deseja excluir esta atividade?")) {
+      const updatedItems = items.filter((item) => item.id !== id);
+      setItems(updatedItems);
+      localStorage.setItem("listItems", JSON.stringify(updatedItems));
+    }
   };
 
   return (
-    <div className="justify-center pl-24 pr-24">
+    <div className="justify-center pl-24 pr-24 mb-12">
       <h1 className="text-4xl mt-12 text-center font-[family-name:var(--font-geist-mono)]">
-        Lista da Atividades
+        Lista de Atividades Cadastradas
       </h1>
-      <div className="flex justify-around pl-24 pr-24 flex-row mt-16 mb-4">
+      <div className="flex justify-between pl-16 pr-8 flex-row mt-16 mb-4">
         <p className="text-2xl font-[family-name:var(--font-geist-mono)]">
           Nome da Atividade
         </p>
         <p className="text-2xl font-[family-name:var(--font-geist-mono)]">
-          Responsável
+          Excluir
         </p>
       </div>
       <ul className="list">
-        <div className="pt-0 pb-4 ml-60 mr-60 rounded-xl border border-solid border-black/[.08] dark:border-white/[.145]">
+        <div className="pt-1 self-center bg-black p-6 rounded-lg">
           {items.map((item) => (
             <li key={item.id} className="list-item">
-              <a href="lista/detalhes" className="flex justify-between rounded-lg border dark:border-white/[.145] pl-12 pr-12 text-xl mt-6 font-[family-name:var(--font-geist-mono)]">
-                <p>{item.name}</p>
-                <p>{item.resp}</p>
-              </a>
+              <div className="flex pt-2 justify-between h-12 bg-stone-900 rounded-lg border dark:border-white/[.145] pl-12 pr-12 text-xl mt-6 font-[family-name:var(--font-geist-mono)]">
+                <a target="_blank" href={`lista/detalhes?id=${item.id}`}>
+                  <p>{item.name}</p>
+                </a>
+                <button className="mb-2" onClick={() => handleDelete(item.id)}>
+                  <Image
+                    src="/delete.png"
+                    alt="Icone Lista"
+                    width={25}
+                    height={25}
+                  />
+                </button>
+              </div>
             </li>
           ))}
         </div>
